Day 8: Remove incorrect node loop detection in part 1
It may be possible (in general, not with the input data I was given)
that we only reach ZZZ in part 1 after looping over all left/right
instructions at least once. Thus, it is not enough to check whether we
have followed a node back to AAA (the starting node), since this does
not mean we are stuck in a loop. To correctly check if we are in a
node loop we would need to something like what is done in part 2.

Here, for simplicity, we will just assume that the return condition is
eventually met.