Day 5: Replace Futures with parallel streams and tidy up
Using parallel streams in part 1 of the problem subjectively does not
speed up the calculation much. The refactoring done here is mostly to
change the for-loop style into a stream API style, which makes the
function body more similar to part 2's solution.

In part 2, the use of ExecutorService and Futures is replaced with
parallel streams. Previously, the program ran for about 3:45 minutes on
my machine. With the outher stream parallelized it's about 20 seconds
faster, and more than an additional minute faster with parallelization
on the inner streams. This takes the runtime down to about 2:20 minutes,
total.