Day 7: Tidy-up and prepare structure for switching Jacks to Jokers
Tidy-up some of the variable and class names, as well as small other
code tweaks.

Introduces (admittedly ugly) way of switching Jacks to Jokers for part 2
of this puzzle by passing a boolean `useJokersInsteadOfJacks` around.
In the future, this will need to be cleaned up and done properly. But
for now, this will get us closer to solving part 2.